requirements.txt
streamlit
requests
beautifulsoup4
sentence-transformers
chromadb
llama-index
groq
python-dotenv
